Update hnn_qt5.py
This includes functionality for stimulating at the dendrite and saving the layer voltages, and it also fixes the units. NOTE: the viewing calcium and compartment voltages functionality is NOT workable for multiple cells yet! (It will run but not work per say) However I think the units update is nice and maybe important.