Fix/beautify units
There were a couple of unit discrepancies that I had fixed in the code in my own HNN but it might be good to update this in the GUI when we create the new version with parameter optimization. You can double check them: it seems the HCN and T-type calcium should be in S/cm² instead of pS/µm², and the capacitive density should be in µF/cm² instead of F/cm². I also think it looks better to have "µm²" instead of "micron2" and things like that.